feat: add components to build up the landing page
The components comprise of hero to represent the hero section, about to represent the about us section, services to represent the services section, contact to represent the contact section and footer to represent the footer section. The components are then individually imported in the landing page component and arranged as preferred.